+++ pytorch_lightning/trainer/model_hooks.py
@@ -6,8 +6,9 @@ from pytorch_lightning.core.lightning import LightningModule
 
 class TrainerModelHooksMixin(ABC):
 
-    def is_function_implemented(self, f_name):
-        model = self.get_model()
+    def is_function_implemented(self, f_name, model=None):
+        if model is None:
+            model = self.get_model()
         f_op = getattr(model, f_name, None)
         return callable(f_op)