+++ tests/test_deprecated.py
@@ -44,6 +44,12 @@ def test_tbd_remove_in_v0_10_0_trainer():
     with pytest.deprecated_call(match='v0.10.0'):
         assert trainer.test_percent_check == rnd_val
 
+    trainer = Trainer()
+    with pytest.deprecated_call(match='v0.10.0'):
+        trainer.proc_rank = 0
+    with pytest.deprecated_call(match='v0.10.0'):
+        assert trainer.proc_rank == trainer.global_rank
+
 
 def test_tbd_remove_in_v0_9_0_trainer():
     # test show_progress_bar set by progress_bar_refresh_rate